FYT_algorithm_project/requirements.txt
python-docx
jieba
pandas
pymysql
fasttext
tqdm
numpy
torch
transformers
accelerate
ray
streamlit
scikit-learn
elasticsearch
hanlp
flask
flask_cors
matplotlib
PyYAML
xmindparser
requests
pathos
tensorboard
plotly
ipython
six
regex
colorlog
datasets
cn2an
googletrans
docarray
pycorrector
loguru
networkx
pypinyin
paddle2onnx
keras
autogluon
SQLAlchemy
thulac
h5py
connectorx
onnxruntime
pyltp
Pillow
gevent
fastapi
pydantic
uvicorn
Werkzeug
addressparser
farm-haystack
nltk
tabulate
colorama
joblib
haystack
tabulate